+++ package.json
@@ -50,6 +50,7 @@
         "clsx": "^2.1.1",
         "cmdk": "^1.0.0",
         "dexie": "^4.0.8",
+        "express": "^5.1.0",
         "fast-deep-equal": "^3.1.3",
         "html-to-image": "^1.11.11",
         "i18next": "^23.14.0",
